backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
moviepy>=1.0.3
transformers>=4.31.0
torch>=2.0.0
librosa>=0.10.0
soundfile>=0.12.1
pydantic-settings>=2.0.0
opencv-python-headless>=4.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.0.0
